bb-imager-cli: Make lib
Will allow for better testing.
