+++ bb-imager-cli/src/main.rs
@@ -1,16 +1,8 @@
-mod cli;
-mod helpers;
-
-use bb_flasher::{BBFlasherTarget, DownloadFlashingStatus, LocalImage};
-use clap::{CommandFactory, Parser};
-use cli::{Commands, DestinationsTarget, Opt, TargetCommands};
-use helpers::LocalStringFile;
-use std::path::PathBuf;
-use std::sync::mpsc;
+use clap::Parser;
 use tracing_subscriber::{filter::LevelFilter, layer::SubscriberExt, util::SubscriberInitExt};
 
 fn main() {
-    let opt = Opt::parse();
+    let opt = bb_imager_cli::cli::Opt::parse();
 
     if opt.verbose {
         tracing_subscriber::registry()
@@ -24,636 +16,5 @@ fn main() {
             .expect("Failed to register tracing_subscriber");
     }
 
-    match opt.command {
-        Commands::Flash { target, quiet } => flash(*target, quiet),
-        Commands::Format { dst, quiet } => format(dst, quiet),
-        Commands::ListDestinations {
-            target,
-            no_frills,
-            no_filter,
-        } => list_destinations(target, no_frills, no_filter),
-        Commands::GenerateCompletion { shell } => generate_completion(shell),
-    }
-}
-
-fn flash(target: TargetCommands, quite: bool) {
-    if quite {
-        flash_internal(target, None)
-    } else {
-        std::thread::scope(|s| {
-            let (tx, rx) = mpsc::sync_channel(2);
-
-            s.spawn(move || {
-                let term = console::Term::stdout();
-                let bar_style = indicatif::ProgressStyle::with_template(
-                    "{msg:15}  [{wide_bar}] [{percent:3} %]",
-                )
-                .expect("Failed to create progress bar");
-                let bars = indicatif::MultiProgress::new();
-
-                let mut last_bar: Option<indicatif::ProgressBar> = None;
-                let mut last_state = DownloadFlashingStatus::Preparing;
-                let mut stage = 1;
-
-                // Setting initial stage as Preparing
-                term.write_line(&stage_msg(DownloadFlashingStatus::Preparing, stage))
-                    .unwrap();
-
-                while let Ok(progress) = rx.recv() {
-                    // Skip if no change in stage
-                    if progress == last_state {
-                        continue;
-                    }
-
-                    match (progress, last_state) {
-                        // Take care when just progress needs to be updated
-                        (
-                            DownloadFlashingStatus::DownloadingProgress(p),
-                            DownloadFlashingStatus::DownloadingProgress(_),
-                        )
-                        | (
-                            DownloadFlashingStatus::FlashingProgress(p),
-                            DownloadFlashingStatus::FlashingProgress(_),
-                        ) => {
-                            last_bar.as_ref().unwrap().set_position((p * 100.0) as u64);
-                        }
-                        // Create new bar when stage has changed
-                        (DownloadFlashingStatus::DownloadingProgress(p), _)
-                        | (DownloadFlashingStatus::FlashingProgress(p), _) => {
-                            if let Some(b) = last_bar.take() {
-                                b.finish();
-                            }
-
-                            stage += 1;
-
-                            let temp_bar = bars.add(indicatif::ProgressBar::new(100));
-                            temp_bar.set_style(bar_style.clone());
-                            temp_bar.set_message(stage_msg(progress, stage));
-                            temp_bar.set_position((p * 100.0) as u64);
-                            last_bar = Some(temp_bar);
-                        }
-                        // Print stage when entering a new stage without progress
-                        (DownloadFlashingStatus::Verifying, _)
-                        | (DownloadFlashingStatus::Customizing, _)
-                        | (DownloadFlashingStatus::Preparing, _) => {
-                            if let Some(b) = last_bar.take() {
-                                b.finish();
-                            }
-
-                            stage += 1;
-                            term.write_line(&stage_msg(progress, stage)).unwrap();
-                        }
-                    }
-
-                    last_state = progress;
-                }
-
-                if let Some(b) = last_bar.take() {
-                    b.finish();
-                }
-            });
-
-            flash_internal(target, Some(tx))
-        })
-    }
-    .expect("Failed to flash")
-}
-
-fn flash_internal(
-    target: TargetCommands,
-    chan: Option<mpsc::SyncSender<DownloadFlashingStatus>>,
-) -> anyhow::Result<()> {
-    match target {
-        TargetCommands::Sd {
-            dst,
-            hostname,
-            timezone,
-            keymap,
-            user_name,
-            user_password,
-            wifi_ssid,
-            wifi_password,
-            img,
-            ssh_key,
-            usb_enable_dhcp,
-            bmap,
-            sysconfig,
-            cloud_init,
-            file_destination,
-        } => {
-            // TODO: Remove fallback in the future.
-            if !sysconfig && !cloud_init {
-                tracing::warn!("No config format specified. Using sysconfig by default");
-            }
-
-            let user = user_name.map(|x| (x, user_password.unwrap()));
-            let wifi = wifi_ssid.map(|x| (x, wifi_password.unwrap()));
-
-            let dst = check_macos_device_path(dst);
-
-            let customization = if hostname.is_some()
-                || timezone.is_some()
-                || keymap.is_some()
-                || user.is_some()
-                || wifi.is_some()
-                || ssh_key.is_some()
-                || usb_enable_dhcp
-            {
-                let mut customization = bb_flasher::sd::FlashingSdLinuxConfig::sysconfig(
-                    hostname.clone(),
-                    timezone.clone(),
-                    keymap.clone(),
-                    user.clone(),
-                    wifi.clone(),
-                    ssh_key.clone(),
-                    Some(usb_enable_dhcp),
-                );
-
-                if cloud_init {
-                    customization.extend([bb_flasher::sd::FlashingSdLinuxConfig::cloud_init(
-                        hostname, timezone, keymap, user, wifi, ssh_key,
-                    )]);
-                }
-
-                customization
-            } else {
-                bb_flasher::sd::FlashingSdLinuxConfig::none()
-            };
-
-            tracing::info!("Customization: {:#?}", customization);
-
-            if file_destination {
-                bb_flasher::sd::Flasher::with_file_dest(
-                    LocalImage::new(img).into_image_fn(),
-                    bmap.map(LocalStringFile::new).map(|x| x.into_fn()),
-                    dst,
-                    customization,
-                )
-            } else {
-                bb_flasher::sd::Flasher::new(
-                    LocalImage::new(img).into_image_fn(),
-                    bmap.map(LocalStringFile::new).map(|x| x.into_fn()),
-                    dst.try_into().unwrap(),
-                    customization,
-                )
-            }
-            .flash(chan, None)
-        }
-        TargetCommands::SdBootUpdate { img, dst } => {
-            std::thread::scope(|s| {
-                let tx = if let Some(chan) = chan {
-                    let (tx, rx) = mpsc::sync_channel(4);
-                    s.spawn(move || {
-                        let _ = chan.try_send(DownloadFlashingStatus::Preparing);
-                        while let Ok(msg) = rx.recv() {
-                            // Safeguard for initial rewinds
-                            if msg > 0.01 {
-                                let _ =
-                                    chan.try_send(DownloadFlashingStatus::FlashingProgress(msg));
-                            }
-                        }
-                    });
-                    Some(tx)
-                } else {
-                    None
-                };
-
-                bb_flasher::sd::UpdateBootFlasher::with_file_dest(
-                    LocalImage::new(img).into_archive_fn(tx),
-                    dst,
-                    None,
-                )
-                .flash()
-            })
-        }
-        #[cfg(feature = "bcf_cc1352p7")]
-        TargetCommands::Bcf {
-            img,
-            dst,
-            no_verify,
-        } => bb_flasher::bcf::cc1352p7::Flasher::new(
-            LocalImage::new(img).into_image_fn(),
-            dst.into(),
-            !no_verify,
-            None,
-        )
-        .flash(chan),
-        #[cfg(feature = "bcf_msp430")]
-        TargetCommands::Msp430 { img, dst } => {
-            bb_flasher::bcf::msp430::Flasher::new(LocalImage::new(img).into_image_fn(), dst.into())
-                .flash(chan)
-        }
-        #[cfg(feature = "pb2_mspm0")]
-        TargetCommands::Pb2Mspm0 { no_eeprom, img } => {
-            bb_flasher::pb2::mspm0::Flasher::new(LocalImage::new(img).into_image_fn(), !no_eeprom)
-                .flash(chan)
-        }
-        #[cfg(feature = "dfu")]
-        TargetCommands::Dfu { identifier, imgs } => {
-            if imgs.len() % 2 == 1 {
-                panic!("Failed to parse input images");
-            }
-
-            let img_list = imgs
-                .chunks_exact(2)
-                .map(|x| {
-                    (
-                        x[0].to_string(),
-                        LocalImage::new(PathBuf::from(&x[1]).into()).into_image_fn(),
-                    )
-                })
-                .collect();
-
-            bb_flasher::dfu::Flasher::from_identifier(img_list, &identifier, None)
-                .unwrap()
-                .flash(chan)
-        }
-        #[cfg(all(
-            any(feature = "zepto_uart", feature = "zepto_i2c"),
-            not(target_os = "linux")
-        ))]
-        TargetCommands::Zepto {
-            img,
-            dst,
-            no_verify,
-        } => bb_flasher::mspm0::Flasher::no_prep(
-            LocalImage::new(img).into_image_fn(),
-            dst.into(),
-            !no_verify,
-            None,
-        )
-        .flash(chan),
-        #[cfg(all(
-            any(feature = "zepto_uart", feature = "zepto_i2c"),
-            target_os = "linux"
-        ))]
-        TargetCommands::Zepto {
-            img,
-            dst,
-            no_verify,
-            reset_gpio,
-            bsl_gpio,
-        } => match (reset_gpio, bsl_gpio) {
-            (Some(reset), Some(bsl)) => bb_flasher::mspm0::Flasher::gpio_by_name(
-                LocalImage::new(img).into_image_fn(),
-                dst.into(),
-                !no_verify,
-                None,
-                reset,
-                bsl,
-            )
-            .flash(chan),
-            (None, None) => bb_flasher::mspm0::Flasher::no_prep(
-                LocalImage::new(img).into_image_fn(),
-                dst.into(),
-                !no_verify,
-                None,
-            )
-            .flash(chan),
-            _ => panic!("Invalid arguments"),
-        },
-    }
-}
-
-#[cfg(target_os = "macos")]
-fn check_macos_device_path(dst: PathBuf) -> PathBuf {
-    if dst.to_string_lossy().starts_with("/dev/disk")
-        && !dst.to_string_lossy().starts_with("/dev/rdisk")
-    {
-        let rdisk = dst.to_string_lossy().replace("/dev/disk", "/dev/rdisk");
-        if std::path::Path::new(&rdisk).exists() {
-            let term = console::Term::stderr();
-            let _ = term.write_line(&format!(
-                "{} You are using a buffered device path: {}\n\
-                 {} For significantly faster flashing, use the raw device path: {}\n",
-                console::style("Warning:").yellow().bold(),
-                dst.display(),
-                console::style("Tip:").green().bold(),
-                rdisk
-            ));
-
-            let _ = term.write_str(&format!(
-                "Do you want to switch to {}? [Y/n] ",
-                console::style(&rdisk).bold()
-            ));
-
-            // Simple stdin read since we don't have dialoguer
-            let mut input = String::new();
-            std::io::stdin()
-                .read_line(&mut input)
-                .expect("Failed to read line");
-
-            let input = input.trim().to_lowercase();
-            if input.is_empty() || input == "y" || input == "yes" {
-                let _ = term.write_line(&format!("Switching to {}\n", rdisk));
-                return PathBuf::from(rdisk);
-            }
-        }
-    }
-
-    dst
-}
-
-#[cfg(not(target_os = "macos"))]
-fn check_macos_device_path(dst: PathBuf) -> PathBuf {
-    dst
-}
-
-fn format(dst: PathBuf, quite: bool) {
-    let term = console::Term::stdout();
-
-    let config = bb_flasher::sd::FormatFlasher::new(dst.try_into().unwrap());
-    config.flash().unwrap();
-
-    if !quite {
-        term.write_line("Formatting successful").unwrap();
-    }
-}
-
-fn no_frills_list_destinations<T: BBFlasherTarget + Send + 'static>(no_filter: bool) {
-    let term = console::Term::stdout();
-    let dsts = T::destinations(!no_filter);
-
-    for d in dsts {
-        term.write_line(&d.identifier()).unwrap();
-    }
-}
-
-fn list_destinations(target: DestinationsTarget, no_frills: bool, no_filter: bool) {
-    if no_frills {
-        match target {
-            DestinationsTarget::Sd => {
-                no_frills_list_destinations::<bb_flasher::sd::Target>(no_filter)
-            }
-            #[cfg(feature = "dfu")]
-            DestinationsTarget::Dfu => {
-                no_frills_list_destinations::<bb_flasher::dfu::Target>(no_filter)
-            }
-            #[cfg(feature = "bcf_cc1352p7")]
-            DestinationsTarget::Bcf => {
-                no_frills_list_destinations::<bb_flasher::bcf::cc1352p7::Target>(no_filter)
-            }
-            #[cfg(feature = "bcf_msp430")]
-            DestinationsTarget::Msp430 => {
-                no_frills_list_destinations::<bb_flasher::bcf::msp430::Target>(no_filter)
-            }
-            #[cfg(feature = "pb2_mspm0")]
-            DestinationsTarget::Pb2Mspm0 => {
-                no_frills_list_destinations::<bb_flasher::pb2::mspm0::Target>(no_filter)
-            }
-            #[cfg(any(feature = "zepto_uart", feature = "zepto_i2c"))]
-            DestinationsTarget::Zepto => {
-                no_frills_list_destinations::<bb_flasher::mspm0::Target>(no_filter)
-            }
-        }
-        return;
-    }
-
-    let term = console::Term::stdout();
-
-    match target {
-        DestinationsTarget::Sd => {
-            const NAME_HEADER: &str = "SD Card";
-            const PATH_HEADER: &str = "Path";
-            const SIZE_HEADER: &str = "Size (in G)";
-            const BYTES_IN_GB: u64 = 1024 * 1024 * 1024;
-
-            let dsts_str: Vec<_> = bb_flasher::sd::Target::destinations(!no_filter)
-                .into_iter()
-                .map(|x| {
-                    (
-                        x.to_string().trim().to_string(),
-                        x.identifier().to_string(),
-                        (x.size() / BYTES_IN_GB).to_string(),
-                    )
-                })
-                .collect();
-
-            let max_name_len = dsts_str
-                .iter()
-                .map(|x| x.0.len())
-                .chain([NAME_HEADER.len()])
-                .max()
-                .unwrap();
-            let max_path_len = dsts_str
-                .iter()
-                .map(|x| x.1.len())
-                .chain([PATH_HEADER.len()])
-                .max()
-                .unwrap();
-            let max_size_len = dsts_str
-                .iter()
-                .map(|x| x.2.len())
-                .chain([SIZE_HEADER.len()])
-                .max()
-                .unwrap();
-
-            let table_border = format!(
-                "+-{}-+-{}-+-{}-+",
-                std::iter::repeat_n('-', max_name_len).collect::<String>(),
-                std::iter::repeat_n('-', max_path_len).collect::<String>(),
-                std::iter::repeat_n('-', SIZE_HEADER.len()).collect::<String>(),
-            );
-
-            term.write_line(&table_border).unwrap();
-
-            term.write_line(&format!(
-                "| {} | {} | {: <6} |",
-                console::pad_str(NAME_HEADER, max_name_len, console::Alignment::Left, None),
-                console::pad_str(PATH_HEADER, max_path_len, console::Alignment::Left, None),
-                console::pad_str(SIZE_HEADER, max_size_len, console::Alignment::Left, None),
-            ))
-            .unwrap();
-
-            term.write_line(&table_border).unwrap();
-
-            for d in dsts_str {
-                term.write_line(&format!(
-                    "| {} | {} | {} |",
-                    console::pad_str(&d.0, max_name_len, console::Alignment::Left, None),
-                    console::pad_str(&d.1, max_path_len, console::Alignment::Left, None),
-                    console::pad_str(&d.2, max_size_len, console::Alignment::Right, None)
-                ))
-                .unwrap();
-            }
-
-            term.write_line(&table_border).unwrap();
-        }
-        #[cfg(feature = "dfu")]
-        DestinationsTarget::Dfu => {
-            const NAME_HEADER: &str = "Device";
-            const BUS_NUMBER_HEADER: &str = "Bus Number";
-            const ADDRESS_HEADER: &str = "Addresss";
-            const VENDOR_ID_HEADER: &str = "Vendor Id";
-            const PRODUCT_ID_HEADER: &str = "Product Id";
-
-            let dsts_str: Vec<_> = bb_flasher::dfu::Target::destinations(!no_filter)
-                .into_iter()
-                .map(|x| {
-                    (
-                        x.to_string().trim().to_string(),
-                        format!("{:#04x}", x.bus_number()),
-                        format!("{:#04x}", x.port_num()),
-                        format!("{:#06x}", x.vendor_id()),
-                        format!("{:#06x}", x.product_id()),
-                    )
-                })
-                .collect();
-
-            let max_name_len = dsts_str
-                .iter()
-                .map(|x| x.0.len())
-                .chain([NAME_HEADER.len()])
-                .max()
-                .unwrap();
-
-            let table_border = format!(
-                "+-{}-+-{}-+-{}-+-{}-+-{}-+",
-                std::iter::repeat_n('-', max_name_len).collect::<String>(),
-                std::iter::repeat_n('-', BUS_NUMBER_HEADER.len()).collect::<String>(),
-                std::iter::repeat_n('-', ADDRESS_HEADER.len()).collect::<String>(),
-                std::iter::repeat_n('-', VENDOR_ID_HEADER.len()).collect::<String>(),
-                std::iter::repeat_n('-', PRODUCT_ID_HEADER.len()).collect::<String>(),
-            );
-
-            term.write_line(&table_border).unwrap();
-
-            term.write_line(&format!(
-                "| {} | {} | {} | {} | {} |",
-                console::pad_str(NAME_HEADER, max_name_len, console::Alignment::Left, None),
-                BUS_NUMBER_HEADER,
-                ADDRESS_HEADER,
-                VENDOR_ID_HEADER,
-                PRODUCT_ID_HEADER,
-            ))
-            .unwrap();
-
-            term.write_line(&table_border).unwrap();
-
-            for d in dsts_str {
-                term.write_line(&format!(
-                    "| {} | {} | {} | {} | {} |",
-                    console::pad_str(&d.0, max_name_len, console::Alignment::Left, None),
-                    console::pad_str(
-                        &d.1,
-                        BUS_NUMBER_HEADER.len(),
-                        console::Alignment::Right,
-                        None
-                    ),
-                    console::pad_str(&d.2, ADDRESS_HEADER.len(), console::Alignment::Right, None),
-                    console::pad_str(
-                        &d.3,
-                        VENDOR_ID_HEADER.len(),
-                        console::Alignment::Right,
-                        None
-                    ),
-                    console::pad_str(
-                        &d.4,
-                        PRODUCT_ID_HEADER.len(),
-                        console::Alignment::Right,
-                        None
-                    ),
-                ))
-                .unwrap();
-            }
-
-            term.write_line(&table_border).unwrap();
-        }
-        #[cfg(feature = "bcf_msp430")]
-        DestinationsTarget::Msp430 => {
-            no_frills_list_destinations::<bb_flasher::bcf::msp430::Target>(no_filter)
-        }
-        #[cfg(feature = "bcf_cc1352p7")]
-        DestinationsTarget::Bcf => {
-            no_frills_list_destinations::<bb_flasher::bcf::cc1352p7::Target>(no_filter)
-        }
-        #[cfg(feature = "pb2_mspm0")]
-        DestinationsTarget::Pb2Mspm0 => {
-            no_frills_list_destinations::<bb_flasher::pb2::mspm0::Target>(no_filter)
-        }
-        #[cfg(any(feature = "zepto_uart", feature = "zepto_i2c"))]
-        DestinationsTarget::Zepto => {
-            no_frills_list_destinations::<bb_flasher::mspm0::Target>(no_filter)
-        }
-    }
-}
-
-const fn progress_msg(status: DownloadFlashingStatus) -> &'static str {
-    match status {
-        DownloadFlashingStatus::Preparing => "Preparing  ",
-        DownloadFlashingStatus::DownloadingProgress(_) => "Downloading",
-        DownloadFlashingStatus::FlashingProgress(_) => "Flashing",
-        DownloadFlashingStatus::Verifying => "Verifying",
-        DownloadFlashingStatus::Customizing => "Customizing",
-    }
-}
-
-fn stage_msg(status: DownloadFlashingStatus, stage: usize) -> String {
-    format!("[{stage}] {}", progress_msg(status))
-}
-
-fn generate_completion(target: clap_complete::Shell) {
-    let mut cmd = Opt::command();
-    const BIN_NAME: &str = env!("CARGO_PKG_NAME");
-
-    clap_complete::generate(target, &mut cmd, BIN_NAME, &mut std::io::stdout())
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn progress_msg_maps_each_status() {
-        assert_eq!(progress_msg(DownloadFlashingStatus::Preparing), "Preparing  ");
-        assert_eq!(
-            progress_msg(DownloadFlashingStatus::DownloadingProgress(0.5)),
-            "Downloading"
-        );
-        assert_eq!(
-            progress_msg(DownloadFlashingStatus::FlashingProgress(0.5)),
-            "Flashing"
-        );
-        assert_eq!(progress_msg(DownloadFlashingStatus::Verifying), "Verifying");
-        assert_eq!(
-            progress_msg(DownloadFlashingStatus::Customizing),
-            "Customizing"
-        );
-    }
-
-    #[test]
-    fn stage_msg_prefixes_stage_number() {
-        assert_eq!(
-            stage_msg(DownloadFlashingStatus::Preparing, 3),
-            "[3] Preparing  "
-        );
-        assert_eq!(
-            stage_msg(DownloadFlashingStatus::Verifying, 1),
-            "[1] Verifying"
-        );
-    }
-
-    #[test]
-    fn local_string_file_reads_contents() {
-        let dir = tempfile::tempdir().unwrap();
-        let path = dir.path().join("bmap.txt");
-        std::fs::write(&path, "bmap-contents").unwrap();
-
-        let read = LocalStringFile::new(path.into_boxed_path()).into_fn();
-        assert_eq!(&*read().unwrap(), "bmap-contents");
-    }
-
-    #[test]
-    fn local_string_file_missing_path_errors() {
-        let read =
-            LocalStringFile::new(PathBuf::from("/nonexistent/bmap.txt").into_boxed_path()).into_fn();
-        assert!(read().is_err());
-    }
-
-    #[cfg(not(target_os = "macos"))]
-    #[test]
-    fn check_macos_device_path_is_identity_off_macos() {
-        let dst = PathBuf::from("/dev/disk2");
-        assert_eq!(check_macos_device_path(dst.clone()), dst);
-    }
+    bb_imager_cli::run(opt);
 }
